move components into subdirectory
