+++ src/Components/Publishing/Series/Components/SeriesTitle.tsx
@@ -1,8 +1,8 @@
 import React, { Component  } from "react"
 import styled, { StyledFunction } from "styled-components"
-import { pMedia } from "../../Helpers"
-import { Fonts } from "../Fonts"
-import { PartnerBlock } from '../Partner/PartnerBlock'
+import { pMedia } from "../../../Helpers"
+import { Fonts } from "../../Fonts"
+import { PartnerBlock } from '../../Partner/PartnerBlock'
 
 interface Props {
   color?: string,
